유저들의 평가 상세 페이지 컴포넌트 분리 및 데이터 적용방식 수정
